move env to compat
