+++ xmake/core/base/compat/env.lua
